Refactored bugapp into a package instead of struct
The struct was only ever used as a namespace for receiver
functions, a package makes more sense and should be more
maintainable
