+++ bugapp/Edit.go
@@ -1,0 +1,38 @@
+package bugapp
+
+import (
+	"fmt"
+	"github.com/driusan/bug/bugs"
+	"io/ioutil"
+	"log"
+	"os"
+	"os/exec"
+	"strconv"
+)
+
+func Edit(args ArgumentList) {
+	issues, _ := ioutil.ReadDir(string(bugs.GetIssuesDir()))
+
+	// No parameters, print a list of all bugs
+	if len(args) == 1 {
+		idx, err := strconv.Atoi(args[0])
+		if idx > len(issues) || idx < 1 {
+			fmt.Printf("Invalid issue number %d\n", idx)
+			return
+		}
+		dir := bugs.Directory(bugs.GetIssuesDir()) + bugs.Directory(issues[idx-1].Name())
+		cmd := exec.Command(getEditor(), string(dir)+"/Description")
+
+		cmd.Stdin = os.Stdin
+		cmd.Stdout = os.Stdout
+		cmd.Stderr = os.Stderr
+
+		err = cmd.Run()
+		if err != nil {
+			log.Fatal(err)
+		}
+	} else {
+		fmt.Printf("Usage: %s edit issuenum\n", os.Args[0])
+		fmt.Printf("\nNo issue number specified\n")
+	}
+}
