+++ bugapp/Purge.go
@@ -1,0 +1,22 @@
+package bugapp
+
+import (
+	"fmt"
+	"github.com/driusan/bug/bugs"
+	"github.com/driusan/bug/scm"
+)
+
+func Purge() {
+	scm, _, err := scm.DetectSCM()
+
+	if err != nil {
+		fmt.Printf("Error: %s\n", err.Error())
+		return
+	}
+
+	err = scm.Purge(bugs.GetIssuesDir())
+	if err != nil {
+		fmt.Printf("Error: %s\n", err.Error())
+		return
+	}
+}
