+++ bugapp/Pwd.go
@@ -1,0 +1,11 @@
+package bugapp
+
+import (
+    "fmt"
+	"github.com/driusan/bug/bugs"
+)
+
+
+func Pwd() {
+	fmt.Printf("%s", bugs.GetIssuesDir())
+}
