+++ main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"fmt"
 	"os"
-	//"regex"
+	"github.com/driusan/bug/bugapp"
 	"github.com/driusan/bug/bugs"
 )
 
@@ -18,7 +18,6 @@ func getEditor() string {
 }
 
 func main() {
-	app := BugApplication{}
 	if bugs.GetRootDir() == "" {
 		fmt.Printf("Could not find issues directory.\n")
 		fmt.Printf("Make sure either the PMIT environment variable is set, or a parent directory of your working directory has an issues folder.\n")
@@ -28,41 +27,41 @@ func main() {
 	if len(os.Args) > 1 {
 		switch os.Args[1] {
 		case "add", "new", "create":
-			app.Create(os.Args[2:])
+			bugapp.Create(os.Args[2:])
 		case "view", "list":
-			app.List(os.Args[2:])
+			bugapp.List(os.Args[2:])
 		case "priority":
-			app.Priority(os.Args[2:])
+			bugapp.Priority(os.Args[2:])
 		case "status":
-			app.Status(os.Args[2:])
+			bugapp.Status(os.Args[2:])
 		case "milestone":
-			app.Milestone(os.Args[2:])
+			bugapp.Milestone(os.Args[2:])
 		case "tag":
-			app.Tag(os.Args[2:])
+			bugapp.Tag(os.Args[2:])
 		case "mv", "rename", "retitle", "relabel":
-			app.Relabel(os.Args[2:])
+			bugapp.Relabel(os.Args[2:])
 		case "purge":
-			app.Purge()
+			bugapp.Purge()
 		case "rm", "close":
-			app.Close(os.Args[2:])
+			bugapp.Close(os.Args[2:])
 		case "edit":
-			app.Edit(os.Args[2:])
+			bugapp.Edit(os.Args[2:])
 		case "--version", "version":
-			app.Version()
+			bugapp.Version()
 		case "env":
-			app.Env()
+			bugapp.Env()
 		case "dir", "pwd":
-			app.Pwd()
+			bugapp.Pwd()
 		case "commit":
-			app.Commit()
+			bugapp.Commit()
 		case "roadmap":
-			app.Roadmap(os.Args[2:])
+			bugapp.Roadmap(os.Args[2:])
 		case "help":
 			fallthrough
 		default:
-			app.Help(os.Args[1:]...)
+			bugapp.Help(os.Args[1:]...)
 		}
 	} else {
-		app.Help()
+		bugapp.Help()
 	}
 }
